zoltar_backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=1.4
alembic>=1.7
# psycopg2-binary>=2.9 # Use this for PostgreSQL
python-dotenv>=0.20
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
email-validator>=2.0
python-multipart>=0.0.5
python-dateutil>=2.8
apscheduler>=3.10,<4.0
requests>=2.20.0
msal>=1.0.0
starlette>=0.25.0
google-generativeai
apns2

# Production requirements
gunicorn # WSGI/ASGI server
psycopg2-binary # PostgreSQL driver 